+++ services/resetPassword.js
@@ -1,9 +1,5 @@
 import { UserModel } from "../models/User.js";
-import crypto from "crypto";
-import nodemailer from "nodemailer";
-import { NODEMAILER_PASSWORD,NODEMAILER_EMAIL, CLIENT_URL } from "../config/consts.js";
-
-const resetTokens = new Map();
+import { generateResetToken, sendResetPasswordEmailToUser, getResetTokens } from "../services/mails/resetPassword.js";
 
 export default async function sendResetPasswordEmail(req, res) {
   const { email } = req.body;
@@ -15,32 +11,8 @@ export default async function sendResetPasswordEmail(req, res) {
       return res.status(404).json({ message: "Email doesn't exist in the system." });
     }
 
-    const token = crypto.randomBytes(32).toString("hex");
-    resetTokens.set(token, { userId: user._id, createdAt: Date.now() });
-
-    setTimeout(() => {
-      resetTokens.delete(token);
-    }, 1000*60*10); 
-
-    const resetLink = `${CLIENT_URL}/login/setNewPassword?token=${token}`;
-    const transporter = nodemailer.createTransport({
-      service: "gmail",
-      auth: {
-        user: NODEMAILER_EMAIL, 
-        pass: NODEMAILER_PASSWORD, 
-      },
-    });
-
-    await transporter.sendMail({
-      from: `"Reset Password" <${NODEMAILER_EMAIL}>`,
-      to: email,
-      subject: "Reset Your Password",
-      html: `
-        <p>Hello ${user.name},</p>
-        <p>Click the link below to reset your password:</p>
-        <a href="${resetLink}">${resetLink}</a>
-      `
-    });
+    const token = generateResetToken(user._id);
+    await sendResetPasswordEmailToUser(user, token);
 
     return res.status(200).json({ message: "Reset link sent. Please check your Email" });
   } catch (err) {
@@ -49,4 +21,4 @@ export default async function sendResetPasswordEmail(req, res) {
   }
 }
 
-export { resetTokens };
+export const resetTokens = getResetTokens();
